Version 2025.01

+++ src/context/api/userProfile/infraestructure/http-api/delete-userprofile-by-id/delete-by-id.controller.ts
@@ -1,16 +1,21 @@
-import { ClassSerializerInterceptor, Controller, Delete, Param, ParseIntPipe, UseGuards, UseInterceptors } from "@nestjs/common";
+import { BadRequestException, ClassSerializerInterceptor, Controller, Delete, Param, ParseIntPipe, UseGuards, UseInterceptors } from "@nestjs/common";
 import { DeleteByIdHttpDto } from "./delete-by-id-http-dto";
-import { ApiParam, ApiTags } from "@nestjs/swagger";
-import { GetInformationByToken } from "src/context/services/get-information.decorator";
+import { ApiBearerAuth, ApiParam, ApiTags } from "@nestjs/swagger";
+import { GetInformationByToken } from "src/context/decorators/get-information.decorator";
 import { DeleteUserProfileByIdUseCase } from "../../../application/delete-userprofile-by-id-use-case/delete-userprofile-by-id-use-case";
 import { JwtAuthGuard } from "src/context/guards/jwt.guard";
+import { PermissionsGuard } from "src/context/guards/permissions.guard";
+import { Permissions } from "src/context/decorators/permissions.decorator";
+import { PermissionEnum } from "src/context/api/permission/domain/permission.enum";
+import { PrefixEndpointType, PrivateEndpoints } from "src/context/routes/routing";
+import { isMaster } from "src/context/api/userType/domain/user-type.enum";
 
-@ApiTags('user-profile')
-@Controller('user-profile')
+@ApiTags(PrefixEndpointType.PRIVATE)
+@Controller()
 @UseInterceptors(ClassSerializerInterceptor)
 export class DeleteUserProfileByIdController {
 
-    constructor(private readonly deleteUserProfileByIdUseCase: DeleteUserProfileByIdUseCase) {}
+    constructor(private readonly deleteUserProfileByIdUseCase: DeleteUserProfileByIdUseCase) { }
 
     /**
      * GIBT: Get Information By Token
@@ -21,12 +26,20 @@ export class DeleteUserProfileByIdController {
         required: true,
         description: 'The id of the userProfile',
     })
-    @UseGuards(JwtAuthGuard)
-    @Delete('/delete/:id')
+    @UseGuards(JwtAuthGuard, PermissionsGuard)
+    @Permissions(PermissionEnum.DELETE_PROFILE)
+    @Delete(PrivateEndpoints.DELETE_PROFILE)
+    @ApiBearerAuth()
     async run(@Param('id', ParseIntPipe) id: number, @GetInformationByToken() gibt: any): Promise<void> {
-        console.log(gibt.username);
-        const params = new DeleteByIdHttpDto();
-        params.id = id;
-        await this.deleteUserProfileByIdUseCase.execute(params);
+        try {
+            if(id === 1 || isMaster(gibt.user_type)) throw new BadRequestException('You cannot delete this profile');
+            
+            const params = new DeleteByIdHttpDto();
+            params.id = id;
+            await this.deleteUserProfileByIdUseCase.execute(params);
+        }
+        catch (error) {
+            throw new BadRequestException(error.message);
+        }
     }
 }
